feat(utils): add loadMonaco utility function
plus make sure _editor is initially undefined and add checks across code to validate that it has to have a value before doing something to it
