+++ src/platform/code-editor/code-editor.utils.ts
@@ -18,6 +18,37 @@ export function waitUntilMonacoReady(): Observable<void> {
   return monacoReady$.asObservable();
 }
 
+/**
+ * Check if monaco has been loaded
+ */
 export function isMonacoLoaded(): boolean {
   return typeof((<any>window).monaco) === 'object';
+}
+
+/**
+ * Loads monaco
+ */
+export function loadMonaco(): void {
+  // check if the script tag has been created in case another code component has done this already
+  if (!document.getElementById('monaco-loader-script')) {
+    let onGotAmdLoader: any = () => {
+      // Load monaco
+      (<any>window).require.config({ paths: { 'vs': 'assets/monaco/vs' } });
+      (<any>window).require(['vs/editor/editor.main'], () => {
+        // TODO
+      });
+    };
+
+    // Load AMD loader if necessary
+    if (!(<any>window).require) {
+      let loaderScript: HTMLScriptElement = document.createElement('script');
+      loaderScript.id = 'monaco-loader-script';
+      loaderScript.type = 'text/javascript';
+      loaderScript.src = 'assets/monaco/vs/loader.js';
+      loaderScript.addEventListener('load', onGotAmdLoader);
+      document.body.appendChild(loaderScript);
+    } else {
+      onGotAmdLoader();
+    }
+  }
 }
